requirements.txt

Pillow==9.5.0
numpy==1.25.0
python-dotenv==1.0.0
pandas==2.0.3
joblib==1.3.1
fastapi==0.100.1
uvicorn==0.23.1
pydantic==2.1.1
orange3==3.35.0
